Look for bitcode for library input files in other places
If an input file is not bitcode and looks like a library file, treat it
as library so we can look for it in other places.

In conjuction with the previous change, now we can just drop, eg,
libfreetype.dylib bitcode in emscripten/system/lib and "dynamic" linking
would succeed.

Perhaps this can be changed to only lookup in emscripten/system/lib
instead of all lib_dirs, but this is a good start.