Implement support for passing real pointers to glVertexAttribPointer
glVertexAttribPointer can both take a raw pointer as its argument or an
offset into a bound offer.  The WebGL equivalent always relies on a
bound buffer.  This patch enables the first usage of this API on top of
WebGL.  The way we implement this is we record all of the vertex
attribute calls, and then in glDrawElements, we look at all of the
recorded items, and for each one, bind an array buffer, set its data,
and then call the WebGL vertexAttribPointer function with an offset of
0.  That call will operate on the bound buffer.  Once we've gone through
all of the recorded vertex attributes, we proceed with calling the WebGL
drawElements API.

Note that glDrawElements also expects to take a raw pointer, so we also
need to create and bind a new element array buffer, and then set the raw
data passed to us using that buffer, and then use that buffer with an
offset of 0 in the call to WebGL drawElements function.