Improve linking
When calling llvm-ld, if we are compiling a shared library,
just create the bitcode named as the requested target.  This
is like before.

When linking a final binary, create the bitcode named target.bc,
and create and retain normal link output named after target.
When called this way, llvm-ld creates a small bash script that
interprets the bitcode.  Net result is that binaries generated
can be run.  Morever, the bitcode file is not called .bc, easier
to pass to emcc later.

That said, trying to run any bitcode using 'lli' crashes for me.