Adding LLVM intrinsic llvm_uadd_with_overflow_i16.
With -O1, sometimes LLVM generates code which calls its intrinsic
llvm_uadd_with_overflow_i16, which was previously missing in
emscripten's library.

This commit adds that intrinsic, and also adds a test which
causes LLVM to call that intrinsic (only in 'o1'). This test
previously failed in 'o1' but now passes.