Work around a Windows issue that causes 'WindowsError: [Error 6] The handle is invalid' exception when invoking llvm to detect whether bitcode files are valid or invalid. This allows cygwin make 3.80 and mingw32-make to be used to build makefiles that invoke the emcc compiler.
On Windows, if any of the standard output handles are PIPEd, apparently all of the must be. This commit alters all invokations of Popen in all of the emcc files except tests/runner.py that use PIPE in Popen by PIPEing all stdin, stdout and stderr where possible, to workaround the issue. Whether this issue occurs depends on which 'make' tool is used to build Makefiles that invoke emscripten. In particular, the cygwin make 3.80 and mingw32-make that support windows file paths suffer from this problem. Cygwin make 3.81 and 3.82 did not manifest this problem, but they are unusable since they are "deliberately" broken on windows because they dropped support for windows path names.