Fixes to emcc to make autoconf enable shared objs
When using emcc as a linker, autoconf disables
shared object support.  This is because it expects
the text returned by the -v and --help options to
match certain things that GNU ld prints.

Specifically it expects the output of emcc -v to include
the word 'GNU' and the output of emcc -v to match the
regexp ': supported targets.* elf'

Modified the messages returned by emcc for these options
to make autoconf happy.