Fix SDL canvas colors. Closes #1287.
The hello_world_sdl.cpp test seemed to be constructed wrongly; the old
code actually produced a line in the box when run through native clang.
This commit fixes that. I also added an `extern "C"` specifier, which
seems necessary to get the test to compile natively on OS X.

I also discovered that a blank screen was being colored entirely red,
instead of black. This has been fixed and a test added.