Reduce memory usage when assigning registers in a block.
Rather than pre-calculating liveness states and then processing nodes
in execution order, we process nodes in reverse execution order and
use the liveness info implicit in this traversal.  Names are assigned
a register when they're used for the first time, and the register is
freed when we reach an assignment.