Disable emptyasm_aue, as inline asm is now diagnosed with an error.
This testcase was originally inspired by libgcrypt. Compiling
asm("":::"memory") to a no-op would be doable, and would help the code
compile without error, but properly supporting what libgrypt is actually
doing here, being paranoid enough to make repeated stores to the same
memory location to clear it out, would be much harder, because we'd
really need a way to tell the JS engine what's going on so that it doesn't
optimize anything away.