Extract helper functions from main link loop.
This enables future changes which will re-scan archives for objects
(either grouped with -Wl,--start-group, -Wl,--end-group or some new
emcc specific flag).

This is mostly moving lines around and adding some comments. The only
functional change is that if you specify the same archive multiple times
(a workaround for the lack of grouping flags), we will now only extract
it once instead of multiple times. But it will still be scanned multiple
times.