Let emscripten use response files for everything
I think the individual CMAKE_*_ARCHIVE_CREATE and CMAKE_*_ARCHIVE_APPEND directives are not necessary anymore. I think they used to be necessary because CMAKE_AR was not set correctly as a cache variable before. 
The CMAKE_*_CREATE_STATIC_LIBRARY calls are required to force cmake to use response files for objects. This is necessary because cmake has a hard coded limit for the command line which will not work well for windows because emscripten uses bat files to call the compiler  and bat files have a maximum command line length that is smaller than the general limit for windows executables.