Make no-prefixing of library functions work.
This makes it so that functions in JS library files can start
with '$' to signal that they shouldn't be prefixed with an '_' and
be visible to user code.

This is used in the conversion of emval and embind to JS library
format so that their internals are not exposed inappropriately.