implement clock_gettime(CLOCK_MONOTONIC)
It's bad to pretend to support every clock - we should correctly return -1 with EINVAL for the clocks we don't implement. It's also nice to give implementations a way to detect if they have access to a monotonic clock. For comparison, old Linux machines (RHEL 3 for example) don't have CLOCK_MONOTONIC, new ones do.