Switch to a newer unofficial simd.js polyfill
Switch to the new polyfill from
https://github.com/sunfishcode/ecmascript_simd/blob/development/src/ecmascript_simd.js
for now. This anticipates several changes which are expected to be
accepted upstream.

Among other things, this contains the shuffle->swizzle rename and the
shuffleMix->shuffle rename and generalization.

I also added a workaround for the fact that SpiderMonkey isn't allowing
the polyfill to detect whether signMask is implemented, and isn't
allowing it to be overridden.