Override CMake CheckTypeSize.
The CMake CheckTypeSize module is used by many projects to get sizes of types
for the target platform. Its standard implementation builds a file that
contains the type size with the try_compile command, then uses file(STRINGS on
the built function to extract the size from the contents of the built binary.

Since emscripten produces javascript and html instead of binaries, the string
is not embedded directly in the 'binary' as expected.  Alternatively, we use a
try_run and return the size of the type in the return value.