Fix emscripten_get_gamepad_status causing TypeError
Passing the index of a previously disconnected gamepad to the function
emscripten_get_gamepad_status() caused a TypeError. This was possible if
multiple gamepads were connected and one with a small index was removed.
In this case the array returned by navigator.getGamepads() will contain
a null entry. The function emscripten_get_gamepad_status() only checked
for undefined entries and returned EMSCRIPTEN_RESULT_NO_DATA if this was
the case. A null entry was not detected and the null was then passed on
assuming it was an actual gamepad. This eventually caused the TypeError.
The fault was fixed by changing the check to include null (and the other
values evaluating to false which therefore also can not be gamepads).