Provide some tests of eliding stack restoration.
This tests the changes in https://github.com/kripken/emscripten-fastcomp/pull/64

The intent of those changes is to stop restoring the stack in every
single function, regardless of whether or not any stack was used.

other.test_simplify_ifs needed a minor tweak as the generated code is now
better and there are only 2 ifs and not 3 as previous in one of the test
configurations.