Update the POINTER_MASKING feature.
The POINTER_MASKING argument is now a boolean rather than the mask and
indicates whether pointers can be masked to a power-of-two heap
length.

A POINTER_MASKING_OVERFLOW setting has been added to specify the
length added to the TOTAL_MEMORY to allow the compiler to derive that
accesses are within bounds even when adding small constant offsets.

A POINTER_MASKING_DYNAMIC setting has been added, disabled by
default. When disabled, the masking and heap length are baked in at
compile time. When enabled, the masking and heap length are set at
runtime and the POINTER_MASKING_DEFAULT_ENABLED setting sets the
default in this mode.

A heap length declaration function is now emitted for the benefit of
Odin. The masks are emitted using a 'const' rather than a 'var' for
the benefit of Odin.