emcc: Remove EMCC_LEAVE_INPUTS_RAW from env.
If emcc is running with LEAVE_INPUTS_RAW and then launches an
emcc to build something like the struct info, then we don't want
LEAVE_INPUTS_RAW to be active in that emcc subprocess.

This is an issue for people who use emscripten and do their own
bitcode processing and optimization.