Disable the (name+num)+num = name+newnum optimization.
This optimization reassociates the addition and can introduce situations
where intermediate computation wraps around the address space, creating
inefficiency in JS engines which are optimized for no-wraparound cases.

In its place, the fastcomp backend is being enhanced to fold
getelementptrs with constant bases, which is the main case where the
above optimization happens in practice.