Add support for CMAKE_CROSSCOMPILING_EMULATOR.
This variable was added in recent CMake Git master.  It should point to an
executable that acts like an emulator for the target system when
cross-compiling.  For emscripten, this is nodejs.

This executable is used by CMake to run executables in try_run commands.  By
adding support for this variable, it prevents the need to manually populate
the tedious TryRunResults.cmake file.

The emscripten configured path to NodeJS is added to the emcmake command.  In
the Emscripten.cmake file, if CMAKE_CROSSCOMPILING_EMULATOR is not defined
(which is would be if using emcmake), and attempt is made to find nodejs and
use it.