Verify integrity of memory initializer using CRC-32
This is almost the standard CRC-32 algorithm, except that we omit the final
XOR with -1 so that we can easily compare the result against zero.
The length of the initializer is included in the data so that we don't have
to worry about leading zeros (after XOR with the init value of -1).

Useful read: http://www.ross.net/crc/download/crc_v3.txt