Use hex escapes for non-ASCII bytes
This makes the resulting literals more independent from the character
encoding the environment assumes for the resulting file.
It requires slightly more memory, but large bytes are far less common than
small bytes (zero in particular), so the cost should not be too much.
If we want to, we can still make this optional later on.