Rename Settings.CLOSURE_COMPILER.
This renames Settings.CLOSURE_COMPILER and the 'closure' variable
in emcc. Prior to this, it was confusing to have two different
variables with the same name: shared.CLOSURE_COMPILER and
shared.Settings.CLOSURE_COMPILER.

Settings.CLOSURE_COMPILER is now Settings.USE_CLOSURE_COMPILER
to make the intent clear and to disambiguate it from
shared.CLOSURE_COMPILER which is a path to the actual Closure
compiler.