[optimizer] Find MSBuild on 32 bit Windows.
This allows the optimizer to get built on 32 bit Windows. It is still
not building a working binary.

On 32 bit Windows, there isn't a ProgramFiles(x86) environment
variable, so the checks for that fail with a runtime error that
breaks the build.  Instead, check to see if the environment
variable exists and if not, carry on.