Prevent Building.link from forcing absolute paths
In some cases, we have paths to objects files that we need to link
which contain spaces. Since Building.link cannot create a response file
that includes paths with spaces, it will add the object files to the
command line. When we have too many object files, the command line gets
too long, and we see this error on Windows:

WindowsError: [Error 87] The parameter is incorrect

We can work around this issue by giving Building.link relative paths to
the object files (really just file names), and setting the working
directory to by the path with the space.

However, Building.link was unnecessarily converting these relative
paths to absolute paths.

This change prevents Building.link from making that conversion when it
calls the linker. Instead it now uses whatever relative paths it was
given.