Implement getpeername and fix getsockname #3997
The existing getsockname was actually returning the address of the other
end of the connection, which is getpeername. getpeername can fail, if
the destination address is not set ('connect' or 'accept' has never been
called).

getsockname returns the address of the local end of the connection, and
cannot fail; an unbound socket will return '0.0.0.0:0'.

However, at this point a socket that was auto-bound (i.e. has not had
'bind' called on it) is also returning '0.0.0.0:0' for getsockname,
which is incorrect but at least consistently-behaved now.