Make BINARYEN_ROOT a config file setting
Currently the BINARYEN -s setting is used both as a flag to enable wasm
creation with binaryen, and to specify the path. This change keeps -s
BINARYEN=1 to enable binaryen, but moves the path to a setting called
BINARYEN_ROOT in the config file, alongside LLVM_ROOT and
EMSCRIPTEN_ROOT. It also makes the path include the 'bin' directory for
consistency with those directories.