Support multiple arguments in EMCC_CFLAGS
Currently the way the support for EMCC_CFLAGS is written it passes the entire string wholesale into sys.argv.append() which treats it as one single argument. This change splits the string using the shlex module (which handles quoting etc) and appends the items to sys.argv piecewise using extend().