Workaround for file_packager's use-preload-cache option on iOs
Older versions of safari have a very bad implementation of indexeddb.
( see here: https://www.raymondcamden.com/2014/09/25/IndexedDB-on-iOS-8-Broken-Bad/ )

In the file packager 2 object stores are created and each pack's uid is inserted in it with the pack path as key. This results in the same Primary Keys in both object stores. Safari then removes the previously added object with this primary key, regardless of the object store it is in.

The solution for this is simple: Prefix your keys to avoid they are the same.

Webkit bug: https://bugs.webkit.org/show_bug.cgi?id=137154