Support inline null bytes in UTF8ArrayToString()
Change priority/meaning of optional maxBytes parameter. If the bytewise length is provided, do not interpret null bytes (within or at the end of the buffer) as terminators. This way, embedded null bytes can be supported in the decoding process. Rename maxBytes to sizeInBytes to indicate that the parameter is not evaluated together with the null terminator.