Fix sbrk bug
In function _sbrk, it sets Runtime.dynamicAlloc function to some error
function to prevent reentrance. But it does not set Runtime.dynamicAlloc
to the original dynamicAlloc function anywhere. So, effectively, once
_sbrk is called, we cannot call _malloc ever again, because _malloc
calls Runtime.dynamicAlloc.

In asm.js, this problem did not appear because there are two _malloc
functions in the resulting js code, where one is _malloc from library.js
and the other is _malloc compiled from libc. And all calls to _malloc
called the version in libc, which does not call Runtime.dynamicAlloc.

But in WebAssembly, some functions in js code (like
___cxa_find_matching_catch) calls _malloc and this calls _malloc in
library.js, because there is no other _malloc in this case.
So this fails:
_sbrk(N);
_malloc(M);
where N and M is any number.