Fix inlining bug (#4471)
* Fix inlining bug

Currently INLINING_LIMIT option does not do anything for clang. (And in
opt, emcc only checks whether it is 0 or not, so the limit number given
actually does not matter unless it is 0.)
So we give -O2, clang inlines short functions even if INLINING_LIMIT is
nonzero. This patch fixes this bug.

* Add a test

* Add Heejin Ahn to AUTHORS list