Fix bug in wasm table size calculation
When mapping split/homogenous tables to the single wasm table, the size
counting code didn't handle empty tables. These only occur with -s
RELOCATABLE=1, which results in an additional 'X' signature (which has
all the functions) and empty tables for the other signatures.

This resulted in Module['wasmTableSize'] (which feeds into the size of
the WebAssembly.Table object) not matching the initial size of the table
import declaration in the wasm module. Currently V8 has a
limitation that it throws a LinkError in this case, while SM is happy
with it. In any case, since this mismatch is not intentional, it should
be fixed.