This seems to be a common source of issues when trying to run optimised code with separate .js.mem from Node.js, while not being in the same working directory.
It was possible to workaround this and set search path using memoryInitializerPrefixURL since 7e9e24, but I think it makes sense to have a nice default too, which would be to search for .js.mem file in the same directory as .js itself when running using Node.js.

Added regression test for this behaviour, which also revealed that SpiderMonkey / V8 shells already got this right, so this PR aligns Node.js behaviour to others.