Multiprocessing pool cwd fix (#4959)
The issue with reusing the Multiprocessing pool in the test harness was that the pool was created to a CWD that was deleted in between tests, which would cause all later subprocess spawns from the pool workers to fail because the worker processes had CWD pointing to a nonexisting directory. To fix, root the pool subprocesses to a known safe directory (EMSCRIPTEN_TEMP) that is very unlikely to die throughout the main process run.

Reverts PR #4957 because that addressed a workaround only and had a bug that close_multiprocessing_pool() would not bind to closing the right pool each time if multiple ones had been created.

Also reverts PR #4958 because I don't think TMP=. is a feature we should support. Though tested that TMP=. emcc tests/hello_world.c does build fine without errors after this PR.