Remove obsolete locale preallocating code in newlocale
We previously statically allocated a single locale in the event that
we needed to dynamically allocate one during startup. musl changed their
implementation to statically allocate C_LOCALE and UTF8_LOCALE, so we
already get that behavior.
In addition this was causing invalid locales to get allocated instead of
returning NULL, because the preceding logic changed. That was caught by
other.test_locale_wrong