LEGALIZE_JS_FFI option to disable JS FFI mangling.
For JS/Web platform, calls to JS imports are normally wrapped to
convert i64 to i32 and f32 to f64. Likewise calls from JS into exports
do the inverse wrapping. This change provides an option to disable
that wrapping and use the original types for the call. This is useful
for non-web embeddings.

The following fastcomp options get enabled by this option:
    -emscripten-legalize-javascript-ffi=0
    -emscripten-only-wasm

The following binaryen options get enabled by this option:
    --no-legalize-javascript-ffi
    --wasm-only