Fix glfwGetWindowMonitor() to return whether actually in fullscreen
This is an important discrepency in the HTML5 platform vs native.
Browsers require a "user gesture" to enter fullscreen, therefore
attempting to enter with glfwSetWindowMonitor() can fail. If so, then
glfwGetWindowMonitor() should not return the monitor it tried to set,
since glfw applications use this call to check for fullscreen.