Add stubs for sigsetjmp / siglongjmp
Define `sigsetjmp` / `siglongjmp` to be equivalent to `setjmp` / `longjmp`
if _EMSCRIPTEN_SIGLONGJMP_STUB macro is defined, so calls to these
functions will not result in crash due to undefined function and will
work except for `savesigs` semantics.

This `sigsetjmp` stub does not respect `savesigs` argument, but at the time
of writing this Emscripten does not support signals.