Fix detection of CMake try_compile() invocations.
The test executable is no longer named cmTryCompileExec in recent CMake
releases[1][2]. The exact releases containing these changes can also be
found in the Git web UI of CMake. This patch aims to cover the new
naming scheme, without breaking compatibility with the older one. It
should be noted that on Windows both os.sep and '/' should be
considered. The latter case happens when emcc substitutes the response
file with Unix-style file paths in it for linking the test exectuable.

In addition, EMMAKEN_JUST_CONFIGURE_RECURSE should be respected to avoid
endless recursion.

[1] https://gitlab.kitware.com/cmake/cmake/commit/4fbdce2b79b28566ae2c6708c99af5e0e8e0177b
[2] https://gitlab.kitware.com/cmake/cmake/commit/f3e9eeedf43f783c02b2a2a10fc0e632eaf7cfd0