Allow the browser test runner to use arbitrary ports
Get them from the environment and set them on the instance rather than hardcoding them.
This allows browser tests to be run in parallel when driven by an external script (e.g.
to test firefox and chrome simultaneously), and is a step on the way toward making the
browser tests truly parallel like the core tests.