Exit node.js with error status on unhandled promise rejections (#5948)
Currently node will swallow unhandled promise rejections and continue as
if no error occurred. Today it prints a deprecation warning when that
happens, and in the future it will exit with a non-zero status. That
behavior is what we want in any case, so until then, use the
unhandled promise rejection hook to get behavior to match other shells.